MCP/LangchainMCP/pyproject.toml
[project]
name = "langchainmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=0.3.2",
    "mcp-use>=1.2.5",
    "python-dotenv>=1.1.0",
]
